refactor: use hook
